git-svn-id: svn://first.ijs.si:3691/latino/Latino@57 35089d11-96e9-ae4d-82a0-4b2588c93174

+++ Visualization/EditableBitmap.cs
@@ -158,7 +158,7 @@ namespace Latino.Visualization
             {
                 startOffset=source.startOffset+(stride*viewArea.Y)+(viewArea.X*pixelFormatSize);
                 bitmap = new Bitmap(viewArea.Width, viewArea.Height, stride, source.Bitmap.PixelFormat, 
-                    (IntPtr)(((int)byteArray.bitPtr)+startOffset));
+                    (IntPtr)(((long)byteArray.bitPtr)+startOffset));
             }
             finally
             {
